refactor(client): Extract MainPage logic into dedicated MainPageContext
