+++ aiblog-client/src/components/RepoSetup/InputArea/FetchButton.jsx
@@ -1,9 +1,9 @@
 import { Button, CircularProgress } from '@mui/material';
 import SendIcon from '@mui/icons-material/Send';
-import { useAppContext } from '@/contexts/AppContext';
+import { useMainPageContext } from '@/contexts/MainPageContext';
 
 export function FetchButton() {
-  const { isLoading, repoName } = useAppContext();
+  const { isLoading, repoName, handleSubmit } = useMainPageContext();
 
   return (
     <Button
@@ -12,6 +12,7 @@ export function FetchButton() {
       size="large"
       endIcon={isLoading ? null : <SendIcon />}
       disabled={isLoading || !repoName}
+      onClick={handleSubmit}
       sx={{
         height: '56px',
         flexShrink: 0,
