+++ aiblog-client/src/components/RepoSetup/InputArea/index.jsx
@@ -1,10 +1,10 @@
 import { Box, Collapse } from '@mui/material';
 import { RepoInput } from './RepoInput';
 import { FetchButton } from './FetchButton';
-import { useAppContext } from '@/contexts/AppContext';
+import { useMainPageContext } from '@/contexts/MainPageContext';
 
 export function InputArea() {
-  const { repoName } = useAppContext();
+  const { repoName } = useMainPageContext();
 
   return (
     <Box sx={{ display: 'flex', alignItems: 'flex-start', gap: 2 }}>
